Improving all scripts. Trying to add tests and good stuff like comments.
