+++ requirements.txt
@@ -1,0 +1,5 @@
+numpy
+xarray
+pandas
+matplotlib
+scipy
