+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="statistical_transformations",
+    version="0.0.1",
+    packages=find_packages(),
+    url="https://github.com/dlagrava/fdc_bias_correction",
+    description="Statistical transformations for flow",
+    install_requires=[],
+)
